Add OpenSK changes for Tock v2
This has still some very rough edges, hence, this is considered a WIP
commit. Next to the missing `firmware_protect` driver for TockOs (for
disabling JTAG support and preventing dumping of the firmware), this
includes non-working examples. Additionally, it is important to note
that OpenSK can not be built anymore on x86 host machines, since
`libtock_runtime` (part of libtock-rs 2.0 WIP) is only running on ARM
and riscv hosts. Hence a normal invocation of `cargo build` or `cargo
check` on your average dev machine will fail, but you can still
cross-compile for ARM or RISCV and/or run `cargo check --features std`
to see if the code compiles fine, but you **need** to apply the
*alloc_init* patch for the libtock-rs submodule before running the build
commands.

Change submodules back to upstream